+++ src/App.js
@@ -2,7 +2,7 @@ import { BrowserRouter, Route, Routes } from "react-router-dom";
 import Home from "./pages/Home";
 import Login from "./pages/Login";
 import Singup from "./pages/Signup";
-import FindPW from "./pages/FindPW";
+import FindPW from "./pages/ResetPW";
 import ModifyPW from "./pages/ModifyPW";
 
 function App() {
