+++ src/pages/ResetPW.css
@@ -1,6 +1,6 @@
-/* FindPWPage.css */
+/* ResetPWPage.css */
 
-.FindPW-container {
+.ResetPW-container {
     display: flex;
     justify-content: center;
     align-items: center;
@@ -8,7 +8,7 @@
     background-color: lightgray;
 }
 
-.FindPW-form {
+.ResetPW-form {
     width: center;
     padding: 20px;
     border-radius: 15px;
@@ -19,21 +19,21 @@
     align-items: flex-start; /* 수정: 수직 정렬을 왼쪽으로 변경 */
 }
 
-.FindPW-namebox {
+.ResetPW-namebox {
     width: 300px;
     text-align: left;
     margin-bottom: 5px;
     align-self: flex-start; /* 수정: 왼쪽 정렬 */
 }
 
-.FindPW-idbox {
+.ResetPW-idbox {
     width: 300px;
     text-align: left;
     margin-bottom: 5px;
     align-self: flex-start; /* 수정: 왼쪽 정렬 */
 }
 
-.FindPW-button {
+.ResetPW-button {
     width: 70px;
     padding: 10px;
     margin-left: 20px;
@@ -44,7 +44,7 @@
     cursor: pointer;
 }
 
-.FindPW-returnbutton {
+.ResetPW-returnbutton {
     width: 300px;
     padding: 10px;
     margin-top: 50px;
